move cli utils in hop/__main__.py to hop/utils/cli.py
